Refactor: Improve copy-paste functionality with clipboard API
